More 64bit fixes

+++ FDK/Tools/Programs/makeotf/makeotf_lib/source/cffread/cffread.c
@@ -9,7 +9,7 @@
 #include <stdlib.h>
 #include <limits.h>
 #include <math.h>
-
+#include <stdio.h>
 #include "cffread.h"
 #include "dictops.h"
 
@@ -48,7 +48,7 @@ typedef int32_t Fixed;
 #define FIX2INT(f)    ((short)(((f) + FixedHalf) >> 16))
 #define FIX2DBL(f)    ((double)((f) / fixedScale))
 #define DBL2FIX(d)    ((Fixed)((d) * fixedScale + ((d) < 0 ? -0.5 : 0.5)))
-#define DBL2INT(d)    ((long)((d) + ((d) < 0 ? -0.5 : 0.5)))
+#define DBL2INT(d)    ((int32_t)((d) + ((d) < 0 ? -0.5 : 0.5)))
 #define CEIL(f)        ((Fixed)(((f) + 0xffff) & 0xffff0000))
 #define FLOOR(f)    ((Fixed)((f) & 0xffff0000))
 #define PATH_FUNC_DEFINED(func) (h->path.cb != NULL && h->path.cb->func != NULL)
@@ -86,7 +86,7 @@ typedef struct {                  /* FDArray element */
 typedef union {                   /* Stack element */
 	double d;
 	Fixed f;
-	long l;
+	int32_t l;
 } StkElement;
 
 typedef enum {                   /* Stack element type */
@@ -376,7 +376,7 @@ static void pushFix(cffCtx h, Fixed f) {
 }
 
 /* Push integer number on stack (as fixed) */
-static void pushInt(cffCtx h, long l) {
+static void pushInt(cffCtx h, int32_t l) {
 	if (h->stack.cnt >= h->stack.max) {
 		fatal(h, "stack overflow");
 	}
@@ -385,7 +385,7 @@ static void pushInt(cffCtx h, long l) {
 }
 
 /* Push long number on stack (as long) */
-static void pushLong(cffCtx h, long l) {
+static void pushLong(cffCtx h, int32_t l) {
 	if (h->stack.cnt >= h->stack.max) {
 		fatal(h, "stack overflow");
 	}
@@ -431,7 +431,7 @@ static Fixed popFix(cffCtx h) {
 			return h->stack.array[h->stack.cnt].f;
 
 		case STK_LONG: {
-			unsigned long l = h->stack.array[h->stack.cnt].l;
+			uint32_t l = h->stack.array[h->stack.cnt].l;
 			if (l < INT(FixedMin) || l > INT(FixedMax)) {
 				fatal(h, "range check\n");
 			}
@@ -443,15 +443,15 @@ static Fixed popFix(cffCtx h) {
 	return 0;    /* For quiet compilation */
 }
 
-/* Pop number from stack (return long) */
-static long popInt(cffCtx h) {
+/* Pop number from stack (return int32_t) */
+static int32_t popInt(cffCtx h) {
 	if (h->stack.cnt < 1) {
 		fatal(h, "stack underflow");
 	}
 	switch (h->stack.type[--h->stack.cnt]) {
 		case STK_DOUBLE: {
 			double d = h->stack.array[h->stack.cnt].d;
-			if (d < LONG_MIN || d > LONG_MAX) {
+			if (d < INT32_MIN || d > INT32_MAX) {
 				fatal(h, "range check\n");
 			}
 			else {
@@ -494,14 +494,14 @@ static Fixed indexFix(cffCtx h, int i) {
 }
 
 /* Return indexed stack element (as long) */
-static long indexInt(cffCtx h, int i) {
+static int32_t indexInt(cffCtx h, int i) {
 	if (i < 0 || i >= h->stack.cnt) {
 		fatal(h, "stack check");
 	}
 	switch (h->stack.type[i]) {
 		case STK_DOUBLE: {
 			double d = h->stack.array[i].d;
-			if (d < LONG_MIN || d > LONG_MAX) {
+			if (d < INT32_MIN || d > INT32_MAX) {
 				fatal(h, "range check\n");
 			}
 			else {
@@ -563,17 +563,17 @@ static void seekbyte(cffCtx h, Offset offset) {
 	((unsigned char)(((h)->data.left-- == 0) ? fillbuf(h) : *(h)->data.next++))
 
 /* Get integer of specified size starting at buffer pointer */
-static unsigned long getnum(cffCtx h, int size) {
-	unsigned long value;
+static uint32_t getnum(cffCtx h, int size) {
+	uint32_t value;
 	switch (size) {
 		case 4:
-			value = (unsigned long)GETBYTE(h) << 24;
-			value |= (unsigned long)GETBYTE(h) << 16;
+			value = (uint32_t)GETBYTE(h) << 24;
+			value |= (uint32_t)GETBYTE(h) << 16;
 			value |= GETBYTE(h) << 8;
 			return value | GETBYTE(h);
 
 		case 3:
-			value = (unsigned long)GETBYTE(h) << 16;
+			value = (uint32_t)GETBYTE(h) << 16;
 			value |= GETBYTE(h) << 8;
 			return value | GETBYTE(h);
 
@@ -621,7 +621,7 @@ static Offset INDEXGet(cffCtx h, INDEX *index,
 		fatal(h, "INDEX bounds");    /* Requested element out-of-bounds */
 	}
 	/* Get offset */
-	seekbyte(h, index->offset + (unsigned long)element * index->offSize);
+	seekbyte(h, index->offset + (uint32_t)element * index->offSize);
 	offset = getnum(h, index->offSize);
 
 	/* Compute length */
@@ -974,10 +974,10 @@ static Fixed PMRand(void) {
 #define M 2147483647
 #define Q (M / A)
 #define R (M % A)
-	static long seed = 1;
-	long hi = seed / Q;
-	long lo = seed % Q;
-	long test = A * lo - R * hi;
+	static int32_t seed = 1;
+	int32_t hi = seed / Q;
+	int32_t lo = seed % Q;
+	int32_t test = A * lo - R * hi;
 	seed = (test > 0) ? test : test + M;
 	return seed % FixedOne + 1;
 #undef A
@@ -992,9 +992,9 @@ static Fixed PMRand(void) {
    the algorithm listed there doesn't round up the remainder which is required
    in order to obtain a result that is correct to the nearest bit. */
 static Fixed FixedSqrt(Fixed radicand) {
-	unsigned long radical = 0;
-	unsigned long remHi = 0;
-	unsigned long remLo = radicand;
+	uint32_t radical = 0;
+	uint32_t remHi = 0;
+	uint32_t remLo = radicand;
 	int count = 23;    /* = 15+FracBits/2; adjust for other precisions */
 
 	if (radicand <= 0) {
@@ -1002,7 +1002,7 @@ static Fixed FixedSqrt(Fixed radicand) {
 	}
 
 	do {
-		unsigned long term;
+		uint32_t term;
 
 		/* Get next 2 bits of the radicand */
 		remHi = remHi << 2 | remLo >> 30;
@@ -1651,7 +1651,7 @@ static void t2Read(cffCtx h, Offset offset, int init) {
 
 			case t2_shortint: {
 				/* 2 byte number */
-				long byte1 = GETBYTE(h);
+				int32_t byte1 = GETBYTE(h);
 				pushInt(h, byte1 << 8 | GETBYTE(h));
 			}
 			break;
@@ -1674,9 +1674,9 @@ static void t2Read(cffCtx h, Offset offset, int init) {
 
 			case 255: {
 				/* 5 byte number */
-				long byte1 = GETBYTE(h);
-				long byte2 = GETBYTE(h);
-				long byte3 = GETBYTE(h);
+				int32_t byte1 = GETBYTE(h);
+				int32_t byte2 = GETBYTE(h);
+				int32_t byte3 = GETBYTE(h);
 				pushFix(h, byte1 << 24 | byte2 << 16 | byte3 << 8 | GETBYTE(h));
 			}
 			break;
@@ -2048,16 +2048,16 @@ static void DICTRead(cffCtx h, int length, Offset offset, int enable) {
 
 			case cff_shortint: {
 				/* 2 byte number */
-				long byte1 = GETBYTE(h);
+				int32_t byte1 = GETBYTE(h);
 				pushInt(h, byte1 << 8 | GETBYTE(h));
 			}
 			break;
 
 			case cff_longint: {
 				/* 5 byte number */
-				long byte1 = GETBYTE(h);
-				long byte2 = GETBYTE(h);
-				long byte3 = GETBYTE(h);
+				int32_t byte1 = GETBYTE(h);
+				int32_t byte2 = GETBYTE(h);
+				int32_t byte3 = GETBYTE(h);
 				pushLong(h, byte1 << 24 | byte2 << 16 | byte3 << 8 | GETBYTE(h));
 			}
 			break;
